Refactor DB and pebble
Create a new package called pebble which implements the db interface.
Refactors pebble tests and rename function which creates new DB.
