+++ core/trie/trie.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/NethermindEth/juno/core/crypto"
 	"github.com/NethermindEth/juno/core/felt"
-	"github.com/NethermindEth/juno/db"
+	"github.com/NethermindEth/juno/db/pebble"
 	"github.com/bits-and-blooms/bitset"
 )
 
@@ -52,7 +52,7 @@ func NewTrie(storage Storage, height uint, rootKey *bitset.BitSet) *Trie {
 
 // RunOnTempTrie creates an in-memory Trie of height `height` and runs `do` on that Trie
 func RunOnTempTrie(height uint, do func(*Trie) error) error {
-	db, err := db.NewInMemoryDb()
+	db, err := pebble.NewMem()
 	if err != nil {
 		return err
 	}
