+++ sync/sync_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/NethermindEth/juno/blockchain"
-	"github.com/NethermindEth/juno/db"
+	"github.com/NethermindEth/juno/db/pebble"
 	"github.com/NethermindEth/juno/testsource"
 	"github.com/NethermindEth/juno/utils"
 	"github.com/stretchr/testify/assert"
@@ -39,7 +39,7 @@ func TestSyncBlocks(t *testing.T) {
 	}
 	log := utils.NewNopZapLogger()
 	t.Run("sync multiple blocks in an empty db", func(t *testing.T) {
-		testDB := db.NewTestDb()
+		testDB := pebble.NewMemTest()
 		bc := blockchain.NewBlockchain(testDB, utils.MAINNET)
 		synchronizer := NewSynchronizer(bc, gw, log)
 		ctx, cancel := context.WithCancel(context.Background())
@@ -52,7 +52,7 @@ func TestSyncBlocks(t *testing.T) {
 		testBlockchain(t, bc)
 	})
 	t.Run("sync multiple blocks in a non-empty db", func(t *testing.T) {
-		testDB := db.NewTestDb()
+		testDB := pebble.NewMemTest()
 		bc := blockchain.NewBlockchain(testDB, utils.MAINNET)
 		b0, err := gw.BlockByNumber(context.Background(), 0)
 		assert.NoError(t, err)
